+++ src/main/kotlin/com/booleworks/kjobs/control/Maintenance.kt
@@ -8,7 +8,6 @@ import com.booleworks.kjobs.api.JobPersistence
 import com.booleworks.kjobs.common.getOrElse
 import com.booleworks.kjobs.data.Heartbeat
 import com.booleworks.kjobs.data.Job
-import com.booleworks.kjobs.data.JobResult
 import com.booleworks.kjobs.data.JobStatus
 import com.booleworks.kjobs.data.ifError
 import com.booleworks.kjobs.data.orQuitWith
@@ -53,15 +52,15 @@ object Maintenance {
     /**
      * Checks for jobs in status [JobStatus.RUNNING] belonging to instances which seem to be dead. An instance
      * is assumed to be dead *if it missed to update its heartbeat at least twice*.
-     * If such a job has had less than [maxRestarts] restarts, the job is reset to [JobStatus.CREATED],
+     * If such a job has had less than [maxRestartsPerType] restarts, the job is reset to [JobStatus.CREATED],
      * otherwise the job is set to failure. (The assumption is multiple restarts of the same job may indicate
      * that the job's computation is responsible for the death of the instance.)
      */
     suspend fun restartJobsFromDeadInstances(
         jobPersistence: JobPersistence,
-        specificPersistences: Map<String, DataPersistence<*, *>>,
+        persistencesPerType: Map<String, DataPersistence<*, *>>,
         heartbeatInterval: Duration,
-        maxRestarts: Int,
+        maxRestartsPerType: Map<String, Int>,
     ) {
         val liveInstances =
             jobPersistence.fetchHeartBeats(LocalDateTime.now().minus((heartbeatInterval * HEARTBEAT_TIMEOUT_FACTOR).toJavaDuration())).orQuitWith {
@@ -80,7 +79,7 @@ object Maintenance {
             logger.warn("Detected jobs executed by seemingly dead instances. Dead instances are: ${deadInstances.joinToString()}")
         }
 
-        restartJobs(jobsWithDeadInstances, specificPersistences, maxRestarts, "its executing instance seems to be dead")
+        restartJobs(jobsWithDeadInstances, persistencesPerType, maxRestartsPerType, "its executing instance seems to be dead")
     }
 
     /**
@@ -101,36 +100,43 @@ object Maintenance {
      *
      * This may be useful on startup after a possible restart where the instance name did not change.
      *
-     * If a job was restarted more than [maxRestarts] times, the result is set to failure.
+     * If a job was restarted more than [maxRestartsPerType] times, the result is set to failure.
      */
     suspend fun resetMyRunningJobs(
-        persistence: JobPersistence, myInstanceName: String, persistencesPerType: Map<String, DataPersistence<*, *>>, maxRestarts: Int,
+        persistence: JobPersistence,
+        myInstanceName: String,
+        persistencesPerType: Map<String, DataPersistence<*, *>>,
+        maxRestartsPerType: MutableMap<String, Int>,
     ) {
         val myRunningJobs = persistence.allJobsOfInstance(JobStatus.RUNNING, myInstanceName).orQuitWith {
             logger.error("Failed to fetch jobs: $it")
             return
         }
-        restartJobs(myRunningJobs, persistencesPerType, maxRestarts, "its instance has been restarted")
+        restartJobs(myRunningJobs, persistencesPerType, maxRestartsPerType, "its instance has been restarted")
     }
 
-    private suspend fun restartJobs(jobs: List<Job>, specificPersistences: Map<String, DataPersistence<*, *>>, maxRestarts: Int, hint: String) =
-        jobs.forEach { job ->
-            val dataPersistence = specificPersistences[job.type]!!
-            dataPersistence.dataTransaction {
-                if (job.numRestarts >= maxRestarts) {
-                    logger.debug("Setting job with ID ${job.uuid} to failure because $hint and the maximum number of restarts has been reached")
-                    job.status = JobStatus.FAILURE
-                    job.finishedAt = LocalDateTime.now()
-                    persistOrUpdateResult(job, JobResult.error(job.uuid, "The job was aborted because it exceeded the maximum number of $maxRestarts restarts"))
-                } else {
-                    logger.debug("Restarting job with ID ${job.uuid} because $hint")
-                    job.status = JobStatus.CREATED
-                    job.numRestarts += 1
-                    job.executingInstance = null
-                    job.startedAt = null
-                    job.timeout = null
-                }
-                updateJob(job)
-            }.ifError { logger.error("Updating job failed with: $it") }
-        }
+    private suspend fun restartJobs(
+        jobs: List<Job>, specificPersistences: Map<String, DataPersistence<*, *>>, maxRestartsPerType: Map<String, Int>, hint: String
+    ) = jobs.forEach { job ->
+        restartJob(job, maxRestartsPerType[job.type]!!, hint, specificPersistences[job.type]!!)
+    }
+
+    internal suspend fun restartJob(job: Job, maxRestarts: Int, hint: String, persistence: DataPersistence<*, *>) {
+        persistence.dataTransaction {
+            if (job.numRestarts >= maxRestarts) {
+                logger.debug("Setting job with ID ${job.uuid} to failure because $hint and the maximum number of restarts has been reached")
+                job.status = JobStatus.FAILURE
+                job.finishedAt = LocalDateTime.now()
+                persistOrUpdateFailure(job, "The job was aborted because it exceeded the maximum number of $maxRestarts restarts")
+            } else {
+                logger.debug("Restarting job with ID ${job.uuid} because $hint")
+                job.status = JobStatus.CREATED
+                job.numRestarts += 1
+                job.executingInstance = null
+                job.startedAt = null
+                job.timeout = null
+            }
+            updateJob(job)
+        }.ifError { logger.error("Updating job failed with: $it") }
+    }
 }